chore(general-info): shuffle files for v3 build
