+++ packages/inventory-general-info/src/BiosCard/BiosCard.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import LoadingCard from './LoadingCard';
-import { generalMapper } from './dataMapper';
-import { biosSelector } from './selectors';
+import LoadingCard from '../LoadingCard';
+import { generalMapper } from '../dataMapper';
+import { biosSelector } from '../selectors';
 import DateFormat from '@redhat-cloud-services/frontend-components/DateFormat';
-import { isDate } from './constants';
+import { isDate } from '../constants';
 
 const BiosCard = ({ bios, detailLoaded, handleClick }) => (<LoadingCard
     title="BIOS"
