+++ packages/inventory-general-info/src/BiosCard/index.js
@@ -1,0 +1,2 @@
+export { default as BiosCard } from './BiosCard';
+export { default } from './BiosCard';
