+++ packages/inventory-general-info/src/CollectionCard/CollectionCard.test.js
@@ -4,7 +4,7 @@ import { render } from 'enzyme';
 import toJson from 'enzyme-to-json';
 import CollectionCard from './CollectionCard';
 import configureStore from 'redux-mock-store';
-import { collectInfoTest } from './__mock__/selectors';
+import { collectInfoTest } from '../__mock__/selectors';
 
 describe('CollectionCard', () => {
     let initialState;
