+++ packages/inventory-general-info/src/CollectionCard/index.js
@@ -1,0 +1,1 @@
+export { default } from './CollectionCard';
