+++ packages/inventory-general-info/src/ConfigurationCard/index.js
@@ -1,0 +1,2 @@
+export { default } from './ConfigurationCard';
+export * from './ConfigurationCard';
