+++ packages/inventory-general-info/src/EditButton/EditButton.js
