+++ packages/inventory-general-info/src/EditButton/index.js
@@ -1,0 +1,2 @@
+export { default } from './EditButton';
+export * from './EditButton';
