+++ packages/inventory-general-info/src/GeneralInformation/general-information.scss
