+++ packages/inventory-general-info/src/GeneralInformation/index.js
@@ -1,0 +1,2 @@
+export { default } from './GeneralInformation';
+export * from './GeneralInformation';
