+++ packages/inventory-general-info/src/InfoTable/InfoTable.test.js
