+++ packages/inventory-general-info/src/InfoTable/index.js
@@ -1,0 +1,2 @@
+export { default } from './InfoTable';
+export * from './InfoTable';
