+++ packages/inventory-general-info/src/InfrastructureCard/index.js
@@ -1,0 +1,2 @@
+export { default } from './InfrastructureCard';
+export * from './InfrastructureCard';
