+++ packages/inventory-general-info/src/LoadingCard/LoadingCard.js
