+++ packages/inventory-general-info/src/LoadingCard/LoadingCard.test.js
