+++ packages/inventory-general-info/src/LoadingCard/index.js
@@ -1,0 +1,2 @@
+export { default } from './LoadingCard';
+export * from './LoadingCard';
