+++ packages/inventory-general-info/src/OperatingSystemCard/index.js
@@ -1,0 +1,2 @@
+export { default } from './OperatingSystemCard';
+export * from './OperatingSystemCard';
