+++ packages/inventory-general-info/src/SystemCard/index.js
@@ -1,0 +1,2 @@
+export { default } from './SystemCard';
+export * from './SystemCard';
