+++ packages/inventory-general-info/src/TextInputModal/index.js
@@ -1,0 +1,2 @@
+export { default } from './TextInputModal';
+export * from './TextInputModal';
