+++ packages/inventory-general-info/src/constants/constants.js
