+++ packages/inventory-general-info/src/constants/constants.test.js
