+++ packages/inventory-general-info/src/constants/index.js
@@ -1,0 +1,2 @@
+export { default } from './constants';
+export * from './constants';
