+++ packages/inventory-general-info/src/dataMapper/dataMapper.js
