+++ packages/inventory-general-info/src/dataMapper/dataMapper.test.js
