+++ packages/inventory-general-info/src/dataMapper/index.js
@@ -1,0 +1,2 @@
+export { default } from './dataMapper';
+export * from './dataMapper';
