+++ packages/inventory-general-info/src/redux/index.js
@@ -1,0 +1,5 @@
+export * from './action-types';
+export * from './actions';
+export * from './reducers';
+export { default as systemProfileStore } from './reducers';
+export { default } from './reducers';
