+++ packages/inventory-general-info/src/selectors/index.js
@@ -1,0 +1,2 @@
+export { default } from './selectors';
+export * from './selectors';
