+++ packages/inventory-general-info/src/selectors/selectors.js
