+++ packages/inventory-general-info/src/selectors/selectors.test.js
@@ -14,7 +14,7 @@ import {
     infraTest,
     configTest,
     collectInfoTest
-} from './__mock__/selectors';
+} from '../__mock__/selectors';
 
 it('propertiesSelector should return correct data', () => {
     expect(propertiesSelector(testProperties)).toEqual({
